Created Version 1
LIMITATIONS:
1. When you go to the chapters section, the name of the song must be in "ARTIST - TRACK" format for this to properly work. It doesn't play well with numbers or videos where some songs don't have artists.

2. The process for adding username and client ids is a bit tedious right now, so future versions should probably make that more streamlined.

For a first draft though, I like it. I've wanted to do this for a while because my bookmarks tab is full of songs that I need to convert.

Honestly it might have been faster to just manually convert all the videos into playlists but, you know, we're programmers - Why do something in an hour when you can spend 6 badly automating it?